add prioritized filtering capability

+++ ITableFilter.cs
@@ -4,8 +4,10 @@ namespace Replicator
 {
     public interface ITableFilter
     {
-        // Return null to allow all tables, or a Dictionary<string, ulong> 
-        // with the tables to allow and their priorities.
-        Dictionary<string, ulong> Factory();
+        // Return null to allow all tables, or a Dictionary<string, int> 
+        // with the tables to allow and their priorities. Tables with a
+        // a higher priority will be sent before tables with a lower 
+        // priority value.
+        Dictionary<string, int> Factory();
     }
 }
